Refactor: abstracted out Carbon class into components
